+++ lib/vet360/reference_data/countries_response.rb
@@ -17,25 +17,5 @@ module Vet360
         new(response.status, countries: data)
       end
     end
-
-    class StatesResponse < Response
-      attribute :states, Array[Hash]
-
-      def self.from(response)
-        new(response.status, states: response.body['state_list'])
-      end
-    end
-
-    class ZipcodesResponse < Response
-      attribute :zipcodes, Array[Hash]
-
-      def self.from(response)
-        data = response.body['zip_code5_list'].map do |z|
-          { 'zip_code' => z['zip_code5'] }
-        end
-
-        new(response.status, zipcodes: data)
-      end
-    end
   end
 end
